moving to droidlet/tools
[ghstack-poisoned]
