+++ droidlet/tools/active_vision/candidate_selection/launch_candidate_selection.sh
