+++ droidlet/tools/active_vision/common_utils.py
