+++ droidlet/tools/active_vision/explore/explore.sh
