+++ droidlet/tools/active_vision/explore/launch_explore.sh
